+++ singularity-opac/neutrinos/thermal_distributions_neutrinos.hpp
@@ -13,8 +13,8 @@
 // publicly, and to permit others to do so.
 // ======================================================================
 
-#ifndef OPACITIES_THERMAL_DISTRIBUTIONS
-#define OPACITIES_THERMAL_DISTRIBUTIONS
+#ifndef SINGULARITY_OPAC_NEUTRINOS_THERMAL_DISTRIBUTIONS_NEUTRINOS_
+#define SINGULARITY_OPAC_NEUTRINOS_THERMAL_DISTRIBUTIONS_NEUTRINOS_
 
 #include <cmath>
 
@@ -23,10 +23,13 @@
 #include <singularity-opac/base/radiation_types.hpp>
 
 namespace singularity {
+namespace neutrinos {
 
-template <int NSPECIES> struct PlanckDistribution {
+template <int NSPECIES> struct FermiDiracDistributionNoMu {
   PORTABLE_INLINE_FUNCTION
-  static Real ThermalDistributionOfTNu(const Real temp, const Real nu) {
+  Real ThermalDistributionOfTNu(const Real temp, const Real Ye,
+                                const RadiationType type, const Real nu,
+                                Real *lambda = nullptr) {
     using namespace constants;
     Real x = cgs::HPL * nu / (cgs::KBOL * temp);
     Real Bnu = NSPECIES * (2. * cgs::HPL * nu * nu * nu / (cgs::CL * cgs::CL)) *
@@ -34,31 +37,34 @@ template <int NSPECIES> struct PlanckDistribution {
     return Bnu;
   }
   PORTABLE_INLINE_FUNCTION
-  static Real ThermalDistributionOfT(const Real temp) {
+  Real ThermalDistributionOfT(const Real temp, const Real Ye,
+                              const RadiationType type,
+                              Real *lambda = nullptr) {
     using namespace constants;
     return 8. * std::pow(M_PI, 5) * std::pow(cgs::KBOL, 4) * NSPECIES *
            std::pow(temp, 4) /
            (15. * std::pow(cgs::CL, 2) * std::pow(cgs::HPL, 3));
   }
   PORTABLE_INLINE_FUNCTION
-  static Real ThernalNumberDistribution(const Real temp) {
+  Real ThernalNumberDistribution(const Real temp, const Real Ye,
+                                 const RadiationType type,
+                                 Real *lambda = nullptr) {
     using namespace constants;
     constexpr Real zeta3 = 1.20206;
     return 12. * pow(cgs::KBOL, 3) * M_PI * NSPECIES * pow(temp, 3) * zeta3 /
            (pow(cgs::CL, 2) * pow(cgs::HPL, 3));
   }
+  template <typename Emissivity>
+  PORTABLE_INLINE_FUNCTION Real OpacityFromKirkhoff(
+      Emissivity &J, const Real rho, const Real temp, const Real Ye,
+      const RadiationType type, const Real nu, Real *lambda = nullptr) {
+    Real Bnu = ThermalDistributionOfTNu(temp, nu);
+    Real jnu = J.EmissivityPerNuOmega(type, rho, temp, nu, lambda);
+    return jnu / Bnu;
+  }
 };
 
-template <typename Emissivity, typename ThermalDistribution>
-Real OpacityFromKirkhoff(Emissivity &J, ThermalDistribution &B,
-                         const RadiationType type, const Real rho,
-                         const Real temp, const Real nu,
-                         Real *lambda = nullptr) {
-  Real Bnu = B.ThermalDistributionOfTNu(temp, nu);
-  Real jnu = J.EmissivityPerNuOmega(type, rho, temp, nu, lambda);
-  return jnu / Bnu;
-}
-
+} // namespace neutrinos
 } // namespace singularity
 
-#endif // OPACITIES_THERMAL_DISTRIBUTIONS
+#endif // SINGULARITY_OPAC_NEUTRINOS_THERMAL_DISTRIBUTIONS_NEUTRINOS_
