+++ singularity-opac/neutrinos/tophat_emissivity_neutrinos.hpp
@@ -13,48 +13,58 @@
 // publicly, and to permit others to do so.
 // ======================================================================
 
-#ifndef OPACITIES_GRAY_EMISSIVITY
-#define OPACITIES_GRAY_EMISSIVITY
+#ifndef SINGULARITY_OPAC_NEUTRINOS_TOPHAT_EMISSIVITY_NEUTRINOS_
+#define SINGULARITY_OPAC_NEUTRINOS_TOPHAT_EMISSIVITY_NEUTRINOS_
 
 #include <cassert>
 #include <cmath>
 #include <cstdio>
 
+#include <ports-of-call/portability.hpp>
 #include <singularity-opac/base/opac_error.hpp>
 #include <singularity-opac/base/physical_constants.hpp>
-#include <ports-of-call/portability.hpp>
 
-#include "thermal_distributions.hpp"
+#include "thermal_distributions_neutrinos.hpp"
 
 namespace singularity {
+namespace neutrinos {
 
-template <typename ThermalDistribution> class GrayOpacity {
+// Neutrino tophat emissivity from
+// Miller, Ryan, Dolence (2019). arXiv:1903.09273
+template <typename ThermalDistribution> class TophatEmissivity {
 public:
-  GrayOpacity(const Real kappa) : kappa_(kappa) {}
-  GrayOpacity(const ThermalDistribution &dist, const Real kappa)
-      : dist_(dist), kappa_(kappa) {}
-
-  GrayOpacity GetOnDevice() { return *this; }
+  TophatEmissivity(const Real C, const Real numin, const Real numax)
+      : C_(C), numin_(numin), numax_(numax) {}
+  TophatEmissivity(const ThermalDistribution &dist, const Real C,
+                   const Real numin, const Real numax)
+      : dist_(dist), C_(C), numin_(numin), numax_(numax) {}
+  TophatEmissivity GetOnDevice() { return *this; }
   PORTABLE_INLINE_FUNCTION
   int nlambda() const noexcept { return 1; }
   PORTABLE_INLINE_FUNCTION
   void PrintParams() const noexcept {
-    printf("Gray opacity. kappa = %g\n", kappa_);
+    printf("Tophat emissivity. C, numin, numax = %g, %g, %g\n", C_, numin_,
+           numax_);
   }
   inline void Finalize() noexcept {}
 
+  // TODO(JMM): Does this make sense for the tophat?
   PORTABLE_INLINE_FUNCTION
   Real OpacityPerNu(const RadiationType type, const Real rho, const Real temp,
                     const Real nu, Real *lambda = nullptr) {
-    return OpacityFromKirkhoff(*this, dist_);
+    return dist_.OpacityFromKirkhoff(*this, type, rho, temp, nu, lambda);
   }
 
   PORTABLE_INLINE_FUNCTION
   Real EmissivityPerNuOmega(const RadiationType type, const Real rho,
                             const Real temp, const Real nu,
                             Real *lambda = nullptr) {
-    Real Bnu = dist_.ThermalDistributionOfTNu(temp, nu);
-    return kappa_ * Bnu;
+    Real Ye = lambda[0];
+    if (nu > numin_ && nu < numax_) {
+      return C_ * GetYeF(Ye, type) / (4. * M_PI);
+    } else {
+      return 0.;
+    }
   }
 
   PORTABLE_INLINE_FUNCTION
@@ -67,20 +77,39 @@ public:
   PORTABLE_INLINE_FUNCTION
   Real Emissivity(const RadiationType type, const Real rho, const Real temp,
                   Real *lambda = nullptr) {
-    return kappa_ * dist_.ThermalDistributionOfT(temp);
+    Real Ye = lambda[0];
+    Real Bc = C_ * (numax_ - numin_);
+    Real J = Bc * GetYeF(Ye, type);
+    return J;
   }
 
   PORTABLE_INLINE_FUNCTION
   Real NumberEmissivity(RadiationType type, const Real rho, const Real temp,
                         Real *lambda = nullptr) {
-    return kappa_*dist_.ThermalNumberDistribution(temp);
+    using namespace constants;
+    Real Ye = lambda[0];
+    Real Ac = 1 / (cgs::HPL * rho) * C_ * log(numax_ / numin_);
+    return rho * Ac * GetYeF(Ye, type);
   }
 
 private:
-  Real kappa_; // absorption coefficient. Units of 1/cm
+  PORTABLE_INLINE_FUNCTION
+  Real GetYeF(Real Ye, RadiationType type) {
+    if (type == RadiationType::NU_ELECTRON) {
+      return 2. * Ye;
+    } else if (type == RadiationType::NU_ELECTRON_ANTI) {
+      return 1. - 2. * Ye;
+    } else {
+      return 0.;
+    }
+  }
+  Real C_;
+  Real numin_;
+  Real numax_;
   ThermalDistribution dist_;
 };
 
+} // namespace neutrinos
 } // namespace singularity
 
-#endif //  OPACITIES_GRAY_EMISSIVITY
+#endif // SINGULARITY_OPAC_NEUTRINOS_TOPHAT_EMISSIVITY_NEUTRINOS_
